правки

+++ team/models.py
@@ -26,7 +26,7 @@ class TeamsLeaders(models.Model):
 class Checked(models.Model):
     checkpoint=models.ForeignKey("main.Checkpoint", on_delete=models.CASCADE, blank=False, null=True, verbose_name="Чекпоинт")
     team=models.ForeignKey(Teams, on_delete=models.CASCADE, blank=False, null=False, verbose_name="Команда")
-    score=models.PositiveIntegerField(blank=True, null=True, default=None, verbose_name="Баллы")
+    score=models.FloatField(blank=True, null=True, default=None, verbose_name="Баллы")
     is_came=models.BooleanField(blank=True, null=True, default=None, verbose_name="Команда пришла")
     def __str__(self):
         return self.checkpoint.title
